requirements.txt
twscrape
fastapi
uvicorn
python-dotenv